Added the rss parser code
Added new notification service that can be run in the background
-This is not be called yet, so if you wan to start it, start it some
activity to teest it

The service should setup a system alarm to repeat every hour, our
service should catch this alarm and run the check rss feed
method.
The alarm seems to be set but it is not catching it going off, not sure
why. feel free to get it working :)